jvm05，完成大部分指令，修复第4章潜在bug

+++ jvmgo/ch05/rtda/local_vars.go
@@ -4,7 +4,7 @@ import "math"
 
 type LocalVars []Slot
 
-func newLocalVars(maxLocalVars uint) LocalVars {
+func newLocalVars(maxLocalVars uint16) LocalVars {
 	if maxLocalVars > 0 {
 		return make([]Slot, maxLocalVars)
 	}
